pyproject.toml
[tool.poetry]
name = "servier_data_pipeline"
version = "0.1.0"
description = "Data pipeline for processing medical data with BigQuery and GCS"
authors = ["Salmane KECHKAR <salmane.kechkar.pro@gmail.com>"]
readme = "README.md"
#packages = [{ include = "servier_data_pipeline" }]
#package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
polars = "^1.10.0"
google-cloud-storage = "^2.18.2"
google-cloud-bigquery = "^3.26.0"
pydantic = "^2.9.2"
argparse = "^1.4.0"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"
gcsfs = "^2024.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"

pyarrow = "^17.0.0"
google-auth = "^2.35.0"
[tool.poetry.dev-dependencies]
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.7.0"
pytest = "^8.3.3"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
known_third_party = ["pandas", "pydantic", "google"]
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.scripts]
#data_pipeline = "servier_data_pipeline.main:build_data_pipeline"
